moved mod plugin to coreplugins
