+++ coreplugins/mod.py
@@ -12,8 +12,10 @@ class Plugin(commands.Cog, name="Bot Management Commands"):
         self.bot = bot
         super(commands.Cog).__init__()
         bot.register(self)
-        self.blacklist = bot.blacklist
-        self.greylist = bot.greylist
+        self.blacklist = Blacklist(bot)
+        self.greylist = Greylist(bot)
+        bot.coredata['blacklist'] = self.blacklist
+        bot.coredata['greylist'] = self.greylist
 
     ######
     # Misc commands
@@ -188,3 +190,103 @@ class Plugin(commands.Cog, name="Bot Management Commands"):
             await ctx.send("User removed from greylist.")
         else:
             await ctx.send("User's greylist updated.")
+
+
+class Blacklist(object):
+    """Manage the user banlist for using the bot"""
+    
+    def __init__(self, bot):
+        self.bot = bot
+
+    def add_user(self, user: discord.Member):
+        """Adds user to bot blacklist, returns True if added"""
+        if not self.is_member_on_blacklist(user):
+            Config().blacklist.append(user.id)
+            Config().write_config_file()
+            return True
+        else:
+            return False
+
+    def del_user(self, user:discord.Member):
+        """Deletes user to bot blacklist, returns True if deleted"""
+        if self.is_member_on_blacklist(user):
+            Config().blacklist.remove(user.id)
+            Config().write_config_file()
+            return True
+        else:
+            return False
+
+    def get_blacklist_names(self):
+        """Returns the blacklisted member names"""
+        blacklisted_members = ", ".join([self.bot.get_user(id).name for id in Config().blacklist])
+        return blacklisted_members
+
+    def is_member_on_blacklist(self, user: discord.Member):
+        """Returns if user is on bot blacklist"""
+        return self.is_userid_on_blacklist(user.id)
+
+    def is_userid_on_blacklist(self, userID: int):
+        """Returns if user id is on bot blacklist"""
+        if userID in Config().blacklist:
+            return True
+        else:
+            return False
+
+
+class Greylist(object):
+    """Manage the user greylist for using the bot.
+    Users on greylist can't play the bot provided games on their greylist.
+    """
+    
+    def __init__(self, bot):
+        self.bot = bot
+
+    def add(self, user: discord.Member, game: enums.GreylistGames = None):
+        """Adds the given game to users greylist.
+        If game is None, all games will be added.
+        If user is already on list, the user game list will be updated.
+        If user is new on list, True will be returned, otherwise False.
+        """
+        if game is None:
+            game = enums.GreylistGames.ALL
+        was_added = True
+        if self.is_user_on_greylist(user, game):
+            game = Config().greylist[user.id] | game
+            was_added = False
+        Config().greylist[user.id] = game
+        Config().write_config_file()
+        return was_added
+
+    def remove(self, user:discord.Member, game: enums.GreylistGames = None):
+        """Removes the given game from users greylist.
+        If game is None, all games will be removed.
+        If user is not on list, None will be returned.
+        If user was completely removed, True will be returned, otherwise False.
+        """
+        if game is None:
+            game = enums.GreylistGames.ALL
+        was_removed = None
+        if self.is_user_on_greylist(user, game):
+            Config().greylist[user.id] = Config().greylist[user.id] & ~game
+            was_removed = False
+            if Config().greylist[user.id] is enums.GreylistGames.No_Game:
+                del(Config().greylist[user.id])
+                was_removed = True
+        Config().write_config_file()
+        return was_removed
+
+    def get_greylist_names(self):
+        """Returns the greylisted members"""
+        greylisted_members = ", ".join([self.bot.get_user(id).name for id in Config().greylist])
+        return greylisted_members
+
+    def is_user_on_greylist(self, user:discord.Member, game: enums.GreylistGames):
+        """Returns if user is for the given game on the greylist"""
+        return self.is_userid_on_greylist(user.id, game)
+
+    def is_userid_on_greylist(self, userID: int, game: enums.GreylistGames):
+        """Returns if user id is for the given game on the greylist"""
+        if userID in Config().greylist:
+            if Config().greylist[userID] & game is not 0:
+                return True
+        return False
